+++ xalpha/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.9.3"
+__version__ = "0.9.4"
 __author__ = "refraction-ray"
 __name__ = "xalpha"