+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## Unreleased
+
+## v0.9.4 - 2020.07.02
 ### added
 * record 可以直接处理内存中的 DataFrame 账单。
 * ioconf 增加自定义 key_func, 兼容缓存层对 key 支持不够的情况 （大小写不敏感，特殊字符等）